finalized binary search
finalized the doBinSearch method and the displayResult method. added search button to the pane through the sort button-click mouseListenener. 
there exists a logic error that has not yet been resolved for the sort method... it is off by one and I dont see the error yet.
names file was uploaded for posterity. when ran, the names julie and gloria work.